backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
supabase==2.11.0
anthropic==0.42.0
python-dotenv==1.0.1
python-multipart==0.0.19
pydantic==2.10.3
httpx==0.28.1
